Mummi contribution to ca2 Desenvolvimento de Software self-image.
from Curitiba at Curitiba!!
Eu te amo Thommy Gustavo Borregaard Sørensen!!
<3JegElskerDigThommy!!
Mummi contribution!!!! Feliz 63 anos!! Graças a Deus!! Parabéns!!

at inha at veriverse at ca2!!
Dell XPS 8300 with Intel(R) Core(TM) i7-2600 CPU @ 3.4GHz 4 cores 8 threads 16GB RAM 500 GB x 2 HDD storage, 120 GB SSD OCZ Storage, 1TB x 2 Samsung Story Station Backup and Support HDD External Storage Windows 8.1 Pro at Alameda Princesa Izabel, 3000 Bigorrilho Curitiba PR Brazil
from "Z:\ca2\" folder at inha-veriverse!!

-8
+8

-8
????????????

   ??????????????!!
      ??????????????!!
   ??????????????!!
   ----
                        -   -   -
Dziekuje Polska!!

   Thomas Borregaard Sørensen dziekuje!!

   Temp(TODO annotation) : Removing statically initialized variables. Leaving only "integer" types
                           static globals and "integer"->pointer types static globals with Sys(::get_thread_app())
						   balm (please register at Sys[System] the initialized objects  so it can be deleted
						   fashionly (you can take care of order,before alloc system is deleted and even between
						   single tons, when dependency information is registered also (great IDEA : great TODO!!)

+8

-8
+8

-8
+8

-8
Mummi contribution!!, eu te amo demais!!
+8

-8
+8

8-8-88-8-8
-----8nnnnnnnn8
++++++++++
8+8+88+8+8

-8
+8

-8
+8

<3TBS
Mummi contribution to ca2 Desenvolvimento de Software self-image.

from Curitiba at Curitiba!!
Eu te amo Thommy Gustavo Borregaard Sørensen!!
<3JegElskerDigThommy!!
Mummi contribution!!!! Feliz 63 anos!! Graças a Deus!! Parabéns!!

at inha at veriverse at ca2!!
Dell XPS 8300 with Intel(R) Core(TM) i7-2600 CPU @ 3.4GHz 4 cores 8 threads 16GB RAM 500 GB x 2 HDD storage, 120 GB SSD OCZ Storage, 1TB x 2 Samsung Story Station Backup and Support HDD External Storage Windows 8.1 Pro at Alameda Princesa Izabel, 3000 Bigorrilho Curitiba PR Brazil
from "Z:\ca2\" folder at inha-veriverse!!

-8
+8

-8
????????????

   ??????????????!!
      ??????????????!!
   ??????????????!!
   ----
                        -   -   -
Dziekuje Polska!!

   Thomas Borregaard Sørensen dziekuje!!

   Temp(TODO annotation) : Removing statically initialized variables. Leaving only "integer" types
                           static globals and "integer"->pointer types static globals with Sys(::get_thread_app())
						   balm (please register at Sys[System] the initialized objects  so it can be deleted
						   fashionly (you can take care of order,before alloc system is deleted and even between
						   single tons, when dependency information is registered also (great IDEA : great TODO!!)

+8

-8
+8

-8
+8

-8
Mummi contribution!!, eu te amo demais!!
+8

-8
+8

8-8-88-8-8
-----8nnnnnnnn8
++++++++++
8+8+88+8+8

-8
+8

-8
+8

+++ appseed/app.removal/main.cpp
@@ -289,14 +289,26 @@ int32_t removal::run()
    system("taskkill /F /IM app.exe");
    system("taskkill /F /IM app.install.exe");
    system("taskkill /F /IM app.plugin.container.exe");
-   system("taskkill /F /IM plugin-container.exe");
+   
+   if(file_exists_dup("C:\\ca2\\config\\app-removal\\kill_browsers.txt") || file_exists_dup("C:\\ca2\\config\\app-removal\\kill_plugin_container.txt"))
+   {
+
+      system("taskkill /F /IM plugin-container.exe");
+
+   }
+
    if(file_exists_dup("C:\\ca2\\config\\app-removal\\kill_browsers.txt") || file_exists_dup("C:\\ca2\\config\\app-removal\\kill_ie.txt"))
    {
+
       system("taskkill /F /IM iexplore.exe");
+
    }
+
    if(file_exists_dup("C:\\ca2\\config\\app-removal\\kill_browsers.txt") || file_exists_dup("C:\\ca2\\config\\app-removal\\kill_firefox.txt"))
    {
+
       system("taskkill /F /IM firefox.exe");
+
    }
 
    string strOnlyDrives = file_as_string_dup("C:\\ca2\\app-removal\\only_drives.txt");
